Configuración de entidades y anotaciones de relación. Queda funcionando el alta de Cliente, el alta de su respectiva Cuenta y la creación de registro en la tabla relación.
SI SE PUDO BURRO

+++ src/main/java/com/bbva/wshomebanking/presentation/response/account/AccountResponse.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 @Getter
 @Setter
 public class AccountResponse {
-    private long accountNumber;
+    private int accountNumber;
     private BigDecimal balance;
     private String currency;
     private List<Client> clients;
